set up a relationship between companies...
going to users.  in order to do this and complete the circle js was
having an issue as one needed the other to be defined.  so created
a closure for both so that the whole file would be read and then gql
will resolve and at that point they are both "defined"  working
as expected